Refactor instruction module
- Move the decoding logic for the imm field into
  separate methods - one per instruction format.

- Move sign-extension logic into a separate method,
  to keep it in one place.

- Change the signatures of the imm methods from
  i64 to i32. Since no RISC-V instruction is larger
  than 32 bits in size and no imm field is larger
  than 20 bits, working with 64-bit values seemed
  a bit pointless.

- Minor documentation changes.

+++ src/processor.rs
@@ -255,7 +255,7 @@ impl Processor {
                 // instruction, not on the target instruction. No
                 // instruction-address-misaligned exception is generated for a 
                 // conditional branch that is not taken.
-                if instr.imm().unwrap() % (IALIGN / 8) as i64 != 0 {
+                if instr.imm().unwrap() % (IALIGN / 8) as i32 != 0 {
                     todo!();
                 }
